Tests (#15)
* Created BookServiceTest

* Fixed BookServiceTest

* Write CategoryServiceTest

* Fixed BookServiceTest

* Fixed BookServiceTest and CategoryServiceTest

* Fixed bugs for unit tests and added test to controllers

* Fixed bug "Get all books"

* Category Controller Tests

* Fixed mvn run for tests

* Added BookControllerTest

* Added utils for tests

+++ src/main/java/mate/project/store/service/impl/BookServiceImpl.java
@@ -24,6 +24,7 @@ public class BookServiceImpl implements BookService {
     private final BookMapper bookMapper;
     private final BookSpecificationBuilder bookSpecificationBuilder;
 
+    @Transactional
     @Override
     public BookDto save(CreateBookRequestDto bookRequestDto) {
         Book book = bookMapper.toEntity(bookRequestDto);
@@ -60,6 +61,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
+        if (bookRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("Can't find a book with id " + id);
+        }
         bookRepository.deleteById(id);
     }
